+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move report columns settings into a modal
+- Move report settings into a modal
 
 ## [1.2.2] - 2019-08-30